+++ lib/c/Lib_RandomBuffer.h
@@ -11,11 +11,11 @@
 #define HACL_IS_WINDOWS 0
 #endif
 
-bool read_random_bytes(uint64_t len, uint8_t * buf);
+bool read_random_bytes(uint32_t len, uint8_t * buf);
 void * hacl_aligned_malloc(size_t alignment, size_t size);
 void hacl_aligned_free(void * ptr);
 
-void randombytes(uint8_t * x,uint64_t len);
+void randombytes(uint8_t * x,uint32_t len);
 
 #endif // __HACL_LIB_RANDOMBUFFER
